requirements.txt
PyQt5>=5.15
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0